refactor: myDuties
